fix(neo-migration): ui5/cli v3 changes
