+++ neo-migration/fioriapp/ui5-local.yaml
@@ -1,4 +1,4 @@
-specVersion: "2.4"
+specVersion: "2.6"
 metadata:
   name: manageproductsneo2911c
 type: application
