--- neo-migration/fioriapp/webapp/Component-preload.js
@@ -1,35 +1,0 @@
-//@ui5-bundle manageproductsneo2911c/Component-preload.js
-jQuery.sap.registerPreloadedModules({
-"version":"2.0",
-"modules":{
-	"manageproductsneo2911c/Component.js":function(){sap.ui.define(["sap/ui/core/UIComponent","sap/ui/Device","./model/models","./controller/ErrorHandler"],function(t,e,s,i){"use strict";return t.extend("ns.manageproductsneo.Component",{metadata:{manifest:"json"},init:function(){t.prototype.init.apply(this,arguments);this._oErrorHandler=new i(this);this.setModel(s.createDeviceModel(),"device");this.setModel(s.createCommentsModel(),"productFeedback");this.getRouter().initialize()},destroy:function(){this._oErrorHandler.destroy();t.prototype.destroy.apply(this,arguments)},getContentDensityClass:function(){if(this._sContentDensityClass===undefined){if(document.body.classList.contains("sapUiSizeCozy")||document.body.classList.contains("sapUiSizeCompact")){this._sContentDensityClass=""}else if(!e.support.touch){this._sContentDensityClass="sapUiSizeCompact"}else{this._sContentDensityClass="sapUiSizeCozy"}}return this._sContentDensityClass}})});
-},
-	"manageproductsneo2911c/controller/App.controller.js":function(){sap.ui.define(["./BaseController","sap/ui/model/json/JSONModel"],function(e,t){"use strict";return e.extend("ns.manageproductsneo.controller.App",{onInit:function(){var e,n,o=this.getView().getBusyIndicatorDelay();e=new t({busy:true,delay:0});this.setModel(e,"appView");n=function(){e.setProperty("/busy",false);e.setProperty("/delay",o)};this.getOwnerComponent().getModel().metadataLoaded().then(n);this.getOwnerComponent().getModel().attachMetadataFailed(n);this.getView().addStyleClass(this.getOwnerComponent().getContentDensityClass())}})});
-},
-	"manageproductsneo2911c/controller/BaseController.js":function(){sap.ui.define(["sap/ui/core/mvc/Controller","sap/ui/core/UIComponent","sap/m/library"],function(e,t,r){"use strict";return e.extend("ns.manageproductsneo.controller.BaseController",{getRouter:function(){return t.getRouterFor(this)},getModel:function(e){return this.getView().getModel(e)},setModel:function(e,t){return this.getView().setModel(e,t)},getResourceBundle:function(){return this.getOwnerComponent().getModel("i18n").getResourceBundle()},onShareEmailPress:function(){var e=this.getModel("objectView")||this.getModel("worklistView");r.URLHelper.triggerEmail(null,e.getProperty("/shareSendEmailSubject"),e.getProperty("/shareSendEmailMessage"))}})});
-},
-	"manageproductsneo2911c/controller/ErrorHandler.js":function(){sap.ui.define(["sap/ui/base/Object","sap/m/MessageBox"],function(e,s){"use strict";return e.extend("ns.manageproductsneo.controller.ErrorHandler",{constructor:function(e){this._oResourceBundle=e.getModel("i18n").getResourceBundle();this._oComponent=e;this._oModel=e.getModel();this._bMessageOpen=false;this._sErrorText=this._oResourceBundle.getText("errorText");this._oModel.attachMetadataFailed(function(e){var s=e.getParameters();this._showServiceError(s.response)},this);this._oModel.attachRequestFailed(function(e){var s=e.getParameters();if(s.response.statusCode!=="404"||s.response.statusCode===404&&s.response.responseText.indexOf("Cannot POST")===0){this._showServiceError(s.response)}},this)},_showServiceError:function(e){if(this._bMessageOpen){return}this._bMessageOpen=true;s.error(this._sErrorText,{id:"serviceErrorMessageBox",details:e,styleClass:this._oComponent.getContentDensityClass(),actions:[s.Action.CLOSE],onClose:function(){this._bMessageOpen=false}.bind(this)})}})});
-},
-	"manageproductsneo2911c/controller/NotFound.controller.js":function(){sap.ui.define(["./BaseController"],function(n){"use strict";return n.extend("ns.manageproductsneo.controller.NotFound",{onLinkPressed:function(){this.getRouter().navTo("worklist")}})});
-},
-	"manageproductsneo2911c/controller/Object.controller.js":function(){sap.ui.define(["./BaseController","sap/ui/model/json/JSONModel","sap/ui/core/routing/History","../model/formatter","sap/ui/core/format/DateFormat","sap/ui/model/Filter","sap/ui/model/FilterOperator"],function(e,t,n,o,a,i,r){"use strict";return e.extend("ns.manageproductsneo.controller.Object",{formatter:o,onInit:function(){var e,n=new t({busy:true,delay:0});this.getRouter().getRoute("object").attachPatternMatched(this._onObjectMatched,this);e=this.getView().getBusyIndicatorDelay();this.setModel(n,"objectView");this.getOwnerComponent().getModel().metadataLoaded().then(function(){n.setProperty("/delay",e)})},onNavBack:function(){var e=n.getInstance().getPreviousHash();if(e!==undefined){history.go(-1)}else{this.getRouter().navTo("worklist",{},true)}},_onObjectMatched:function(e){var t=e.getParameter("arguments").objectId;this.getModel().metadataLoaded().then(function(){var e=this.getModel().createKey("Products",{ProductID:t});this._bindView("/"+e)}.bind(this))},_bindView:function(e){var t=this.getModel("objectView"),n=this.getModel();this.getView().bindElement({path:e,events:{change:this._onBindingChange.bind(this),dataRequested:function(){n.metadataLoaded().then(function(){t.setProperty("/busy",true)})},dataReceived:function(){t.setProperty("/busy",false)}}})},_onBindingChange:function(){var e=this.getView(),t=this.getModel("objectView"),n=e.getElementBinding();if(!n.getBoundContext()){this.getRouter().getTargets().display("objectNotFound");return}var o=this.getResourceBundle(),a=e.getBindingContext().getObject(),s=a.ProductID,d=a.ProductName;t.setProperty("/busy",false);t.setProperty("/shareSendEmailSubject",o.getText("shareSendEmailObjectSubject",[s]));t.setProperty("/shareSendEmailMessage",o.getText("shareSendEmailObjectMessage",[d,s,location.href]));var u=this.byId("idCommentsList");var c=u.getBinding("items");c.filter(new i("productID",r.EQ,s))},onPost:function(e){var t=a.getDateTimeInstance({style:"medium"});var n=t.format(new Date);var o=this.getView().getBindingContext().getObject();var i=e.getParameter("value");var r={productID:o.ProductID,type:"Comment",date:n,comment:i};var s=this.getModel("productFeedback");var d=s.getData().productComments;d.push(r);s.setData({productComments:d})}})});
-},
-	"manageproductsneo2911c/controller/View1.controller.js":function(){sap.ui.define(["sap/ui/core/mvc/Controller"],function(n){"use strict";return n.extend("ns.manageproductsneo.controller.View1",{onInit:function(){}})});
-},
-	"manageproductsneo2911c/controller/Worklist.controller.js":function(){sap.ui.define(["./BaseController","sap/ui/model/json/JSONModel","../model/formatter","sap/ui/model/Filter","sap/ui/model/FilterOperator","sap/m/MessageToast","sap/m/MessageBox"],function(e,t,s,o,i,n,r){"use strict";return e.extend("ns.manageproductsneo.controller.Worklist",{formatter:s,onInit:function(){var e,s,n=this.byId("table");s=n.getBusyIndicatorDelay();this._oTable=n;this._aTableSearchState=[];e=new t({worklistTableTitle:this.getResourceBundle().getText("worklistTableTitle"),shareOnJamTitle:this.getResourceBundle().getText("worklistTitle"),shareSendEmailSubject:this.getResourceBundle().getText("shareSendEmailWorklistSubject"),shareSendEmailMessage:this.getResourceBundle().getText("shareSendEmailWorklistMessage",[location.href]),tableNoDataText:this.getResourceBundle().getText("tableNoDataText"),tableBusyDelay:0,inStock:0,shortage:0,outOfStock:0,countAll:0});this.setModel(e,"worklistView");this._mFilters={inStock:[new o("UnitsInStock",i.GT,10)],outOfStock:[new o("UnitsInStock",i.LE,0)],shortage:[new o("UnitsInStock",i.BT,1,10)],all:[]};n.attachEventOnce("updateFinished",function(){e.setProperty("/tableBusyDelay",s)})},onUpdateFinished:function(e){var t,s=e.getSource(),o=this.getModel("worklistView"),i=e.getParameter("total");if(i&&s.getBinding("items").isLengthFinal()){t=this.getResourceBundle().getText("worklistTableTitleCount",[i]);this.getModel().read("/Products/$count",{success:function(e){o.setProperty("/countAll",e)}});this.getModel().read("/Products/$count",{success:function(e){o.setProperty("/inStock",e)},filters:this._mFilters.inStock});this.getModel().read("/Products/$count",{success:function(e){o.setProperty("/outOfStock",e)},filters:this._mFilters.outOfStock});this.getModel().read("/Products/$count",{success:function(e){o.setProperty("/shortage",e)},filters:this._mFilters.shortage})}else{t=this.getResourceBundle().getText("worklistTableTitle")}this.getModel("worklistView").setProperty("/worklistTableTitle",t)},onPress:function(e){this._showObject(e.getSource())},onNavBack:function(){history.go(-1)},onSearch:function(e){if(e.getParameters().refreshButtonPressed){this.onRefresh()}else{var t=[];var s=e.getParameter("query");if(s&&s.length>0){t=[new o("ProductName",i.Contains,s)]}this._applySearch(t)}},onRefresh:function(){var e=this.byId("table");e.getBinding("items").refresh()},_showObject:function(e){this.getRouter().navTo("object",{objectId:e.getBindingContext().getProperty("ProductID")})},_applySearch:function(e){var t=this.byId("table"),s=this.getModel("worklistView");t.getBinding("items").filter(e,"Application");if(e.length!==0){s.setProperty("/tableNoDataText",this.getResourceBundle().getText("worklistNoDataWithSearchText"))}},_showErrorMessage:function(e){r.error(e,{styleClass:this.getOwnerComponent().getContentDensityClass()})},onQuickFilter:function(e){var t=this._oTable.getBinding("items"),s=e.getParameter("selectedKey");t.filter(this._mFilters[s])},_handleUnlistActionResult:function(e,t,s,o){if(s===o){n.show(this.getModel("i18n").getResourceBundle().getText("StockRemovedSuccessMsg",[o]))}},_handleReorderActionResult:function(e,t,s,o){if(s===o){n.show(this.getModel("i18n").getResourceBundle().getText("StockUpdatedSuccessMsg",[o]))}},onUnlistObjects:function(){var e,t,s,o,i;e=this.byId("table").getSelectedItems();if(e.length){for(t=0;t<e.length;t++){o=e[t];i=o.getBindingContext().getProperty("ProductID");s=o.getBindingContext().getPath();this.getModel().remove(s,{success:this._handleUnlistActionResult.bind(this,i,true,t+1,e.length),error:this._handleUnlistActionResult.bind(this,i,false,t+1,e.length)})}}else{this._showErrorMessage(this.getModel("i18n").getResourceBundle().getText("TableSelectProduct"))}},onUpdateStockObjects:function(){var e,t,s,o;e=this.byId("table").getSelectedItems();if(e.length){for(t=0;t<e.length;t++){s=e[t].getBindingContext().getPath();o=e[t].getBindingContext().getObject();o.UnitsInStock+=10;this.getModel().update(s,o,{success:this._handleReorderActionResult.bind(this,o.ProductID,true,t+1,e.length),error:this._handleReorderActionResult.bind(this,o.ProductID,false,t+1,e.length)})}}else{this._showErrorMessage(this.getModel("i18n").getResourceBundle().getText("TableSelectProduct"))}}})});
-},
-	"manageproductsneo2911c/i18n/i18n.properties":'# This is the resource bundle for Manage Products\n\n#XTIT: Application name\nappTitle=Manage Products\n\n#YDES: Application description\nappDescription=My Worklist App\n\n#~~~ Worklist View ~~~~~~~~~~~~~~~~~~~~~~~~~~\n#XTIT: Worklist view title\nworklistViewTitle=Manage ProductsPlural\n\n#XTIT: Worklist page title\nworklistTitle=Manage Products\n\n#XTIT: Table view title\nworklistTableTitle=ProductsPlural\n\n#XTOL: Tooltip for the search field\nworklistSearchTooltip=Enter a Product name or a part of it.\n\n#XBLI: text for a table with no data with filter or search\nworklistNoDataWithSearchText=No matching ProductsPlural found\n\n#XTIT: Table view title with placeholder for the number of items\nworklistTableTitleCount=Products ({0})\n\n#XTIT: The title of the column containing the ProductName of Products\ntableNameColumnTitle=ProductName\n\n#XTIT: The title of the column containing the UnitsInStock and the unit of measure\ntableUnitNumberColumnTitle=UnitsInStock\n\n#XTIT: The title of the column containing Supplier name\nTableSupplierColumnTitle=Supplier\n\n#XTIT: The title of the column containing Price\nTablePriceColumnTitle=Price\n\n#XTIT: The title of the column containing Ordered Units\nTableUnitsOrderedColumnTitle=Units Ordered\n\n#XTIT: The title of the column containing Units in Stock\nTableUnitsInStockColumnTitle=Units in Stock\n\n#XBLI: text for a table with no data\ntableNoDataText=No ProductsPlural are currently available\n\n#XLNK: text for link in \'not found\' pages\nbackToWorklist=Show Manage Products\n\n#XTIT: The title of the products quick filter\nWorklistFilterProductsAll=Products\n\n#XTIT: The title of the out of stock products filter\nWorklistFilterOutOfStock=Out of Stock\n\n#XTIT: The title of the low stock products filter\nWorklistFilterShortage=Shortage\n\n#XTIT: The title of the products in stock filter\nWorklistFilterInStock=Plenty in Stock\n\n#text of the button for Products reordering\nTableProductsReorder=Order\n\n#text for the button for Products unlisting\nTablePorductsUnlist=Remove\n\n#Text for no product selected\nTableNoProductsSelected=No product selected\n\n#Product successfully deleted\nStockRemovedSuccessMsg=Product removed\n\n#Product successfully updated\nStockUpdatedSuccessMsg=Product stock level updated\n\n#~~~ Object View ~~~~~~~~~~~~~~~~~~~~~~~~~~\n#XTIT: Object view title\nobjectViewTitle=Products Details\n\n#XTIT: Object page title\nobjectTitle=Products\n\n\n#XTIT: Label for the ProductName\nProductNameLabel=ProductName\n\n#XTIT: Label for the UnitsInStock\nUnitsInStockLabel=UnitsInStock\n\n#Price per unit text\nObjectPriceTitle=Price\n\n#Discontinued text\nObjectDiscontinuedStatusText=Discontinued\n\n#Supplier tab title\nObjectSupplierTabTitle=Supplier Info\n\n#Supplier company name\nObjectSupplierName=Name\n\n#Supplier contact person name\nObjectSupplierContact=Contact\n\n#Supplier contact address\nObjectSupplierAddress=Address\n\n#Supplier zip code\nObjectSupplierZipcode=ZIP Code\n\n#Supplier city name\nObjectSupplierCity=City\n\n#Supplier country\nObjectSupplierCountry=Country\n\n#Object Product ID text\nObjectProductIdText=Product ID\n\n#Comments tab title\nObjectCommentsTabTitle=Comments\n\n#No comments text\nObjectCommentNoData=No Comments\n\n#~~~ Share Menu Options ~~~~~~~~~~~~~~~~~~~~~~~\n\n#XTIT: Send E-Mail subject\nshareSendEmailWorklistSubject=Email subject PLEASE REPLACE ACCORDING TO YOUR USE CASE\n\n#YMSG: Send E-Mail message\nshareSendEmailWorklistMessage=Email body PLEASE REPLACE ACCORDING TO YOUR USE CASE\\r\\n{0}\n\n#XTIT: Send E-Mail subject\nshareSendEmailObjectSubject=Email subject including object identifier PLEASE REPLACE ACCORDING TO YOUR USE CASE {0}\n\n#YMSG: Send E-Mail message\nshareSendEmailObjectMessage=Email body PLEASE REPLACE ACCORDING TO YOUR USE CASE {0} (id: {1})\\r\\n{2}\n\n\n#~~~ Not Found View ~~~~~~~~~~~~~~~~~~~~~~~\n\n#XTIT: Not found view title\nnotFoundTitle=Not Found\n\n#YMSG: The Products not found text is displayed when there is no Products with this id\nnoObjectFoundText=This Product is not available\n\n#YMSG: The Products not available text is displayed when there is no data when starting the app\nnoObjectsAvailableText=No ProductsPlural are currently available\n\n#YMSG: The not found text is displayed when there was an error loading the resource (404 error)\nnotFoundText=The requested resource was not found\n\n#~~~ Error Handling ~~~~~~~~~~~~~~~~~~~~~~~\n\n#YMSG: Error dialog description\nerrorText=Sorry, a technical error occurred! Please try again later.',
-	"manageproductsneo2911c/localService/mockserver.js":function(){sap.ui.define(["sap/ui/core/util/MockServer","sap/ui/model/json/JSONModel","sap/base/util/UriParameters","sap/base/Log"],function(e,t,r,a){"use strict";var o,i="mycompany/myapp/MyWorklistApp/",n=i+"localService/mockdata";var s={init:function(s){var u=s||{};return new Promise(function(s,c){var p=sap.ui.require.toUrl(i+"manifest.json"),f=new t(p);f.attachRequestCompleted(function(){var t=r.fromQuery(window.location.search),c=sap.ui.require.toUrl(n),p=f.getProperty("/sap.app/dataSources/mainService"),l=sap.ui.require.toUrl(i+p.settings.localUri),d=/.*\/$/.test(p.uri)?p.uri:p.uri+"/";if(!o){o=new e({rootUri:d})}else{o.stop()}e.config({autoRespond:true,autoRespondAfter:u.delay||t.get("serverDelay")||200});o.simulate(l,{sMockdataBaseUrl:c,bGenerateMissingMockData:true});var m=o.getRequests();var v=function(e,t,r){r.response=function(r){r.respond(e,{"Content-Type":"text/plain;charset=utf-8"},t)}};if(u.metadataError||t.get("metadataError")){m.forEach(function(e){if(e.path.toString().indexOf("$metadata")>-1){v(500,"metadata Error",e)}})}var g=u.errorType||t.get("errorType"),y=g==="badRequest"?400:500;if(g){m.forEach(function(e){v(y,g,e)})}o.setRequests(m);o.start();a.info("Running the app with mock data");s()});f.attachRequestFailed(function(){var e="Failed to load application manifest";a.error(e);c(new Error(e))})})},getMockServer:function(){return o}};return s});
-},
-	"manageproductsneo2911c/manifest.json":'{"_version":"1.12.0","sap.app":{"id":"manageproductsneo2911c","type":"application","ach":"ach","i18n":{"bundleUrl":"i18n/i18n.properties","supportedLocales":[""],"fallbackLocale":""},"title":"{{appTitle}}","description":"{{appDescription}}","applicationVersion":{"version":"1.0.0"},"resources":"resources.json","dataSources":{"mainService":{"uri":"/northwind/V2/Northwind/Northwind.svc/","type":"OData","settings":{"odataVersion":"2.0","localUri":"localService/metadata.xml"}}}},"sap.ui":{"technology":"UI5","icons":{"icon":"sap-icon://task","favIcon":"","phone":"","phone@2":"","tablet":"","tablet@2":""},"deviceTypes":{"desktop":true,"tablet":true,"phone":true}},"sap.ui5":{"rootView":{"viewName":"ns.manageproductsneo.view.App","type":"XML","async":true,"id":"app"},"dependencies":{"minUI5Version":"1.60.0","libs":{"sap.ui.core":{},"sap.m":{},"sap.f":{}}},"contentDensities":{"compact":true,"cozy":true},"models":{"i18n":{"type":"sap.ui.model.resource.ResourceModel","settings":{"bundleName":"ns.manageproductsneo.i18n.i18n","supportedLocales":[""],"fallbackLocale":""}},"":{"dataSource":"mainService","preload":true}},"routing":{"config":{"routerClass":"sap.m.routing.Router","viewType":"XML","viewPath":"ns.manageproductsneo.view","controlId":"app","controlAggregation":"pages","bypassed":{"target":["notFound"]},"async":true},"routes":[{"pattern":"","name":"worklist","target":["worklist"]},{"pattern":"Products/{objectId}","name":"object","target":["object"]}],"targets":{"worklist":{"viewName":"Worklist","viewId":"worklist","viewLevel":1,"title":"{i18n>worklistViewTitle}"},"object":{"viewName":"Object","viewId":"object","viewLevel":2,"title":"{i18n>objectViewTitle}"},"objectNotFound":{"viewName":"ObjectNotFound","viewId":"objectNotFound"},"notFound":{"viewName":"NotFound","viewId":"notFound"}}}},"sap.platform.hcp":{"uri":"webapp","_version":"1.1.0"}}',
-	"manageproductsneo2911c/model/formatter.js":function(){sap.ui.define(["sap/ui/core/library"],function(r){"use strict";var e=r.ValueState;return{numberUnit:function(r){if(!r){return""}return parseFloat(r).toFixed(2)},quantityState:function(r){if(r===0){return e.Error}else if(r<=10){return e.Warning}else{return e.Success}}}});
-},
-	"manageproductsneo2911c/model/models.js":function(){sap.ui.define(["sap/ui/model/json/JSONModel","sap/ui/Device"],function(e,n){"use strict";return{createDeviceModel:function(){var t=new e(n);t.setDefaultBindingMode("OneWay");return t},createCommentsModel:function(){return new e({productComments:[]})}}});
-},
-	"manageproductsneo2911c/view/App.view.xml":'<mvc:View\n\tcontrollerName="ns.manageproductsneo.controller.App"\n\tdisplayBlock="true"\n\txmlns="sap.m"\n\txmlns:mvc="sap.ui.core.mvc"><Shell><App id="app"\n\t\t\tbusy="{appView>/busy}"\n\t\t\tbusyIndicatorDelay="{appView>/delay}"/></Shell></mvc:View>',
-	"manageproductsneo2911c/view/NotFound.view.xml":'<mvc:View\n\tcontrollerName="ns.manageproductsneo.controller.NotFound"\n\txmlns="sap.m"\n\txmlns:mvc="sap.ui.core.mvc"><MessagePage\n\t\ttitle="{i18n>notFoundTitle}"\n\t\ttext="{i18n>notFoundText}"\n\t\ticon="sap-icon://document"\n\t\tid="page"\n\t\tdescription=""><customDescription><Link id="link" text="{i18n>backToWorklist}" press=".onLinkPressed" /></customDescription></MessagePage></mvc:View>',
-	"manageproductsneo2911c/view/Object.view.xml":'<mvc:View\n\tcontrollerName="ns.manageproductsneo.controller.Object"\n\txmlns="sap.m"\n\txmlns:mvc="sap.ui.core.mvc"\n\txmlns:semantic="sap.f.semantic"\n\txmlns:form="sap.ui.layout.form"\n\txmlns:l="sap.ui.layout"><semantic:SemanticPage\n\t\tid="page"\n\t\theaderPinnable="false"\n\t\ttoggleHeaderOnTitleClick="false"\n\t\tbusy="{objectView>/busy}"\n\t\tbusyIndicatorDelay="{objectView>/delay}"><semantic:titleHeading><Title text="{ProductName}" /></semantic:titleHeading><semantic:headerContent><FlexBox\n\t\t\t\talignItems="Start"\n\t\t\t\tjustifyContent="SpaceBetween"><Panel backgroundDesign="Transparent"><ObjectAttribute\n\t\t\t\t\t\ttitle="{i18n>ObjectProductIdText}"\n\t\t\t\t\t\ttext="{\n\t\t\t\t\t\t\tpath: \'ProductID\',\n\t\t\t\t\t\t\tformatter: \'.formatter.numberUnit\'}"/><ObjectAttribute\n\t\t\t\t\t\ttitle="{i18n>ObjectPriceTitle}"\n\t\t\t\t\t\ttext="{\n\t\t\t\t\t\t\tpath: \'UnitPrice\',\n\t\t\t\t\t\t\tformatter: \'.formatter.numberUnit\'} EUR"/></Panel><Panel backgroundDesign="Transparent"><ObjectNumber\n\t\t\t\t\t\tid="objectHeader"\n\t\t\t\t\t\tunit="PC"\n\t\t\t\t\t\ttextAlign="End"\n\t\t\t\t\t\tstate="{\n\t\t\t\t\t\t\tpath: \'UnitsInStock\',\n\t\t\t\t\t\t\tformatter: \'.formatter.quantityState\'}"\n\t\t\t\t\t\tnumber="{\n\t\t\t\t\t\t\tpath: \'UnitsInStock\',\n\t\t\t\t\t\t\tformatter: \'.formatter.numberUnit\'}"></ObjectNumber><ObjectStatus\n\t\t\t\t\t\ttext="{i18n>ObjectDiscontinuedStatusText}"\n\t\t\t\t\t\tstate="Error"\n\t\t\t\t\t\tvisible="{path:\'Discontinued\'}"/><ProgressIndicator\n\t\t\t\t\t\twidth="300px"\n\t\t\t\t\t\tpercentValue="{UnitsInStock}"\n\t\t\t\t\t\tdisplayValue="{UnitsInStock}"\n\t\t\t\t\t\tshowValue="true"\n\t\t\t\t\t\tstate="{\n\t\t\t\t\t\t\tpath: \'UnitsInStock\',\n\t\t\t\t\t\t\tformatter: \'.formatter.quantityState\'}" /></Panel></FlexBox></semantic:headerContent><semantic:content><l:VerticalLayout width="100%"><Panel\n\t\t\t\tbackgroundDesign="Transparent"\n\t\t\t\theaderText="{i18n>ObjectSupplierTabTitle}"><content><form:SimpleForm\n\t\t\t\t\t\tminWidth="1024"\n\t\t\t\t\t\tmaxContainerCols="2"\n\t\t\t\t\t\teditable="false"\n\t\t\t\t\t\tlayout="ResponsiveGridLayout"\n\t\t\t\t\t\tlabelSpanL="3"\n\t\t\t\t\t\tlabelSpanM="3"\n\t\t\t\t\t\temptySpanL="4"\n\t\t\t\t\t\temptySpanM="4"\n\t\t\t\t\t\tcolumnsL="1"\n\t\t\t\t\t\tcolumnsM="1"><form:content><Label text="{i18n>ObjectSupplierName}"/><Text text="{Supplier/CompanyName}"/><Label text="{i18n>ObjectSupplierAddress}"/><Text text="{Supplier/Address}"/><Label text="{i18n>ObjectSupplierZipcode} / {i18n>ObjectSupplierCity}"/><Text text="{Supplier/PostalCode} / {Supplier/City}"/><Label text="{i18n>ObjectSupplierCountry}"/><Text text="{Supplier/Country}"/></form:content></form:SimpleForm></content></Panel><Panel\n\t\t\t\t\tbackgroundDesign="Transparent"\n\t\t\t\t\theaderText="{i18n>ObjectCommentsTabTitle}"><content><FeedInput post=".onPost"/><List\n\t\t\t\t\t\t\tid="idCommentsList"\n\t\t\t\t\t\t\tnoDataText="{i18n>ObjectCommentNoData}"\n\t\t\t\t\t\t\tshowSeparators="Inner"\n\t\t\t\t\t\t\titems="{\n\t\t\t\t\t\t\t\tpath: \'productFeedback>/productComments\',\n\t\t\t\t\t\t\t\tsorter: {\n\t\t\t\t\t\t\t\t\tpath: \'date\',\n\t\t\t\t\t\t\t\t\tdescending: true\n\t\t\t\t\t\t\t\t}\n\t\t\t\t\t\t\t}"><FeedListItem\n\t\t\t\t\t\t\t\tinfo="{productFeedback>type}"\n\t\t\t\t\t\t\t\ttext="{productFeedback>comment}"\n\t\t\t\t\t\t\t\ttimestamp="{productFeedback>date}"/></List></content></Panel></l:VerticalLayout></semantic:content><semantic:sendEmailAction><semantic:SendEmailAction id="shareEmail" press=".onShareEmailPress"/></semantic:sendEmailAction></semantic:SemanticPage></mvc:View>',
-	"manageproductsneo2911c/view/ObjectNotFound.view.xml":'<mvc:View\n\tcontrollerName="ns.manageproductsneo.controller.NotFound"\n\txmlns="sap.m"\n\txmlns:mvc="sap.ui.core.mvc"><MessagePage\n\t\ttitle="{i18n>objectTitle}"\n\t\ttext="{i18n>noObjectFoundText}"\n\t\ticon="{sap-icon://product}"\n\t\tdescription=""\n\t\tid="page"><customDescription><Link id="link" text="{i18n>backToWorklist}" press=".onLinkPressed" /></customDescription></MessagePage></mvc:View>',
-	"manageproductsneo2911c/view/View1.view.xml":'<mvc:View controllerName="ns.manageproductsneo.controller.View1" xmlns:mvc="sap.ui.core.mvc" displayBlock="true" xmlns="sap.m"><Shell id="shell"><App id="app"><pages><Page id="page" title="{i18n>title}"><content></content></Page></pages></App></Shell></mvc:View>',
-	"manageproductsneo2911c/view/Worklist.view.xml":'<mvc:View\n\tcontrollerName="ns.manageproductsneo.controller.Worklist"\n\txmlns="sap.m"\n\txmlns:mvc="sap.ui.core.mvc"\n\txmlns:semantic="sap.f.semantic"><semantic:SemanticPage\n\t\tid="page"\n\t\theaderPinnable="false"\n\t\ttoggleHeaderOnTitleClick="false"\n\t\tshowFooter="true"><semantic:titleHeading><Title text="{i18n>worklistTitle}" /></semantic:titleHeading><semantic:headerContent><IconTabBar\n\t\t\t\tid="iconTabBar"\n\t\t\t\tselect=".onQuickFilter"\n\t\t\t\texpandable="false"><items><IconTabFilter\n\t\t\t\t\t\tkey="all"\n\t\t\t\t\t\tshowAll="true"\n\t\t\t\t\t\tcount="{worklistView>/countAll}"\n\t\t\t\t\t\ttext="{i18n>WorklistFilterProductsAll}"/><IconTabSeparator/><IconTabFilter\n\t\t\t\t\t\tkey="inStock"\n\t\t\t\t\t\ticon="sap-icon://message-success"\n\t\t\t\t\t\ticonColor="Positive"\n\t\t\t\t\t\tcount="{worklistView>/inStock}"\n\t\t\t\t\t\ttext="{i18n>WorklistFilterInStock}"/><IconTabFilter\n\t\t\t\t\t\tkey="shortage"\n\t\t\t\t\t\ticon="sap-icon://message-warning"\n\t\t\t\t\t\ticonColor="Critical"\n\t\t\t\t\t\tcount="{worklistView>/shortage}"\n\t\t\t\t\t\ttext="{i18n>WorklistFilterShortage}"/><IconTabFilter\n\t\t\t\t\t\tkey="outOfStock"\n\t\t\t\t\t\ticon="sap-icon://message-error"\n\t\t\t\t\t\ticonColor="Negative"\n\t\t\t\t\t\tcount="{worklistView>/outOfStock}"\n\t\t\t\t\t\ttext="{i18n>WorklistFilterOutOfStock}"/></items><content><Table\n\t\t\t\t\t\tid="table"\n\t\t\t\t\t\tbusyIndicatorDelay="{worklistView>/tableBusyDelay}"\n\t\t\t\t\t\tgrowing="true"\n\t\t\t\t\t\tgrowingScrollToLoad="true"\n\t\t\t\t\t\tnoDataText="{worklistView>/tableNoDataText}"\n\t\t\t\t\t\tupdateFinished=".onUpdateFinished"\n\t\t\t\t\t\twidth="auto"\n\t\t\t\t\t\tmode="MultiSelect"\n\t\t\t\t\t\titems="{\n\t\t\t\t\t\t\tpath: \'/Products\',\n\t\t\t\t\t\t\tsorter: {\n\t\t\t\t\t\t\t\tpath: \'ProductName\',\n\t\t\t\t\t\t\t\tdescending: false\n\t\t\t\t\t\t\t},\n\t\t\t\t\t\t\tparameters: {\n\t\t\t\t\t\t\t\t\'expand\': \'Supplier\'\n\t\t\t\t\t\t\t}\n\t\t\t\t\t\t}"><headerToolbar><Toolbar><Title\n\t\t\t\t\t\t\t\tid="tableHeader"\n\t\t\t\t\t\t\t\ttext="{worklistView>/worklistTableTitle}"/><ToolbarSpacer/><SearchField\n\t\t\t\t\t\t\t\tid="searchField"\n\t\t\t\t\t\t\t\ttooltip="{i18n>worklistSearchTooltip}"\n\t\t\t\t\t\t\t\tsearch=".onSearch"\n\t\t\t\t\t\t\t\twidth="auto"></SearchField></Toolbar></headerToolbar><columns><Column id="nameColumn"><Text\n\t\t\t\t\t\t\t\t\tid="nameColumnTitle"\n\t\t\t\t\t\t\t\t\ttext="{i18n>TableNameColumnTitle}"/></Column><Column\n\t\t\t\t\t\t\t\tid="supplierNameColumn"\n\t\t\t\t\t\t\t\tdemandPopin="false"\n\t\t\t\t\t\t\t\tminScreenWidth="Tablet"><Text text="{i18n>TableSupplierColumnTitle}"/></Column><Column\n\t\t\t\t\t\t\t\tid="unitPriceColumn"\n\t\t\t\t\t\t\t\thAlign="End"\n\t\t\t\t\t\t\t\tdemandPopin="true"\n\t\t\t\t\t\t\t\tminScreenWidth="Tablet"><Text text="{i18n>TablePriceColumnTitle}"/></Column><Column id="unitsOnOrderColumn"\n\t\t\t\t\t\t\t\tdemandPopin="true"\n\t\t\t\t\t\t\t\tminScreenWidth="Tablet"\n\t\t\t\t\t\t\t\thAlign="End"><Text text="{i18n>TableUnitsOrderedColumnTitle}"/></Column><Column\n\t\t\t\t\t\t\t\tid="unitsInStockColumn"\n\t\t\t\t\t\t\t\thAlign="End"><Text text="{i18n>TableUnitsInStockColumnTitle}"/></Column></columns><items><ColumnListItem\n\t\t\t\t\t\t\ttype="Navigation"\n\t\t\t\t\t\t\tpress=".onPress"><cells><ObjectIdentifier\n\t\t\t\t\t\t\t\t\ttitle="{ProductName}"/><Text text = "{Supplier/CompanyName}"/><ObjectNumber\n\t\t\t\t\t\t\t\t\tunit="EUR"\n\t\t\t\t\t\t\t\t\tnumber="{\n\t\t\t\t\t\t\t\t\t\tpath: \'UnitPrice\',\n\t\t\t\t\t\t\t\t\t\tformatter: \'.formatter.numberUnit\'}"/><ObjectNumber\n\t\t\t\t\t\t\t\t\tnumber="{\n\t\t\t\t\t\t\t\t\t\tpath: \'UnitsOnOrder\',\n\t\t\t\t\t\t\t\t\t\tformatter: \'.formatter.numberUnit\'}"\n\t\t\t\t\t\t\t\t\t\tunit="PC"/><ObjectNumber\n\t\t\t\t\t\t\t\t\tnumber="{\n\t\t\t\t\t\t\t\t\t\tpath: \'UnitsInStock\',\n\t\t\t\t\t\t\t\t\t\tformatter: \'.formatter.numberUnit\'}"\n\t\t\t\t\t\t\t\t\tunit="PC"\n\t\t\t\t\t\t\t\t\tstate="{\n\t\t\t\t\t\t\t\t\t\tpath: \'UnitsInStock\',\n\t\t\t\t\t\t\t\t\t\tformatter: \'.formatter.quantityState\'}"/></cells></ColumnListItem></items></Table></content></IconTabBar></semantic:headerContent><semantic:sendEmailAction><semantic:SendEmailAction id="shareEmail" press=".onShareEmailPress"/></semantic:sendEmailAction><semantic:positiveAction><semantic:PositiveAction text="{i18n>TableProductsReorder}" press=".onUpdateStockObjects"/></semantic:positiveAction><semantic:negativeAction><semantic:NegativeAction text="{i18n>TablePorductsUnlist}" press=".onUnlistObjects"/></semantic:negativeAction></semantic:SemanticPage></mvc:View>'
-}});
