+++ neo-migration/fioriapp/webapp/test/flpSandbox.html
@@ -46,7 +46,7 @@
 				"app-tile": {
 					title: "Manage Products",
 					description: "",
-					additionalInformation: "SAPUI5.Component=manageproductsneo2911c",
+					additionalInformation: "SAPUI5.Component=ns.manageproductsneo",
 					applicationType: "URL",
 					url: "../"
                     }
@@ -66,12 +66,12 @@
             data-sap-ui-theme="sap_fiori_3"
             data-sap-ui-compatVersion="edge"
             data-sap-ui-language="en"
-            data-sap-ui-resourceroots='{"manageproductsneo2911c": "../"}'
+            data-sap-ui-resourceroots='{"ns.manageproductsneo": "../"}'
             data-sap-ui-frameOptions="allow"
         >
             // NON-SECURE setting for testing environment
         </script>
-		<script id="locate-reuse-libs" src="../utils/locate-reuse-libs.js"
+		<script id="locate-reuse-libs" src="./locate-reuse-libs.js"
 				data-sap-ui-manifest-uri="../manifest.json">
 		</script>
     </head>
