+++ neo-migration/fioriapp/webapp/test/flpSandboxMockServer.html
@@ -34,13 +34,13 @@
 				}
 			},
 			renderers: {
-				fiori2: {
-					componentData: {
-						config: {
-							search: "hidden"
-						}
-					}
+			    fiori2: {
+				componentData: {
+				    config: {
+					search: "hidden"
+				    }
 				}
+			    }
 			},
 			applications: {
 				"app-tile": {
@@ -68,7 +68,7 @@
 			data-sap-ui-resourceroots='{"manageproductsneo2911c": "../", "manageproductsneo2911c.test": "./"}'
 			data-sap-ui-frameOptions="allow"> // NON-SECURE setting for testing environment
 	</script>
-	<script id="locate-reuse-libs" src="../utils/locate-reuse-libs.js"
+	<script id="locate-reuse-libs" src="./locate-reuse-libs.js"
 			data-sap-ui-manifest-uri="../manifest.json"
 			data-sap-ui-componentName="manageproductsneo2911c"
 			data-sap-ui-use-mockserver="true">
